LayoutTests/http/tests/security/contentSecurityPolicy/strict-dynamic-module-script-expected.txt

CONSOLE MESSAGE: Refused to execute a script because it does not appear in the script-src directive of the Content Security Policy.

PASS All the expected CSP violation reports have been fired.
